add authMiddleware; edit update methods

Approved-by: andrija kevic

+++ src/api/user/userService.ts
@@ -2,6 +2,7 @@ import { ISearchUser, IUserService, UserType } from "./interface";
 import { User } from "./userModel";
 import { AuthService } from "../auth/authService"
 import { Op } from "sequelize";
+import bcrypt from 'bcrypt';
 
 export class UserService implements IUserService {
     private readonly authService: AuthService
@@ -16,22 +17,31 @@ export class UserService implements IUserService {
         return users
     }
 
-    async fetchById(id: number): Promise<User> {
+    async fetchById(id: number, userId: number): Promise<User> {
+        if(id != userId) {
+            throw new Error("Unauthorized")
+        }
+
         const user = await User.findOne({
             where: {
-                id: id
+                id: userId
             }
         })
 
         return user as any
     }
 
-    async fetchByEmail(email: string): Promise<User> {
+    async fetchByEmail(email: string, userId: number): Promise<User> {
         const user = await User.findOne({
             where: {
-                email: email
+                email: email,
+                id: userId
             }
         })
+
+        if(!user){
+            throw new Error("Unauthorized")
+        }
 
         return user as any
     }
@@ -46,10 +56,16 @@ export class UserService implements IUserService {
         return user as User;
     }
 
-    async update(id: number, params: UserType): Promise<[affectedCount: number]> {
-        const passwordHash = await this.authService.hashPassword(params.password)
+    async update(paramsId: number, params: UserType): Promise<User> {
+        if(paramsId !== params.id){
+            throw new Error("Unauthorized")
+        }
+        let passwordHash
+        if(params.password){
+            passwordHash = await this.authService.hashPassword(params.password)
+        }
 
-        const updatedRow = await User.update({
+        await User.update({
             role_id: params.role_id,
             full_name: params.full_name,
             email: params.email, 
@@ -59,17 +75,21 @@ export class UserService implements IUserService {
         },
             {
                 where: {
-                    id: id
+                    id: params.id
                 }
             })
 
-        return updatedRow
+            const newUpdatedRow = await User.findByPk(paramsId)
+            return newUpdatedRow as User   
     }
 
-    async delete(id: number): Promise<number> {
+    async delete(id: number, userId: number): Promise<number> {
+        if(userId !== id){
+            throw new Error("Unauthorized")
+        }
         const deletedRow = await User.destroy({
             where: {
-                id: id
+                id: userId
             }
         })
